Moved Output.kt into a common location
