+++ src/main/kotlin/org/wfanet/measurement/loadtest/common/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@wfa_common_jvm//build/rules_kotlin:defs.bzl", "kt_jvm_library")
+
+package(
+    default_testonly = True,
+    default_visibility = ["//src/main/kotlin/org/wfanet/measurement/loadtest:__subpackages__",],
+)
+
+kt_jvm_library(
+    name = "output",
+    srcs = ["Output.kt"],
+)
