+++ src/main/kotlin/org/wfanet/measurement/loadtest/resourcesetup/BUILD.bazel
@@ -16,7 +16,6 @@ package(
 kt_jvm_library(
     name = "resource_setup",
     srcs = [
-        "Output.kt",
         "ResourceSetup.kt",
     ],
     deps = [
@@ -26,6 +25,7 @@ kt_jvm_library(
         "//src/main/kotlin/org/wfanet/measurement/common/identity:principal_identity",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha:account_authentication_server_interceptor",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha:certificates",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/common:output",
         "//src/main/proto/wfa/measurement/api/v2alpha:account_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:accounts_service_kt_jvm_grpc_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:api_key_kt_jvm_proto",
