+++ src/test/kotlin/org/wfanet/panelmatch/integration/k8s/AbstractPanelMatchCorrectnessTest.kt
@@ -290,8 +290,8 @@ abstract class AbstractPanelMatchCorrectnessTest {
 
   protected suspend fun runTest() {
 
-    logger.info { "Wait 10s before polling for results" }
-    delay(10000)
+    logger.info { "Wait 20s before polling for results" }
+    delay(20000)
 
     PortForwarder(getPod(KINGDOM_PUBLIC_DEPLOYMENT_NAME), SERVER_PORT).use { publicForward ->
       val publicAddress: InetSocketAddress = withContext(Dispatchers.IO) { publicForward.start() }
